requirements.txt
importlib-metadata==4.11.2
lxml==4.8.0
Markdown==3.3.6
Pillow==9.0.1
python-docx==0.8.11
zipp==3.7.0
